Refactor url / location / browser
Rather than put it as a reducer, moslty because I wasn't sure how to access the
browser from one, I'm importing common code between the search and dropdown components.
